add the html content in the platform not in the command
so the command can be used to create generic hits.
